+++ jwst/refpix/reference_pixels.py
@@ -310,6 +310,108 @@ class Dataset():
         else:
             self.input_model.data[integration, group] = self.group.copy()
 
+    def log_parameters(self):
+        """Print out the parameters that are valid for this type of data, and
+        those that aren't
+
+        Parameters
+        ----------
+        input_model : JWST datamodel
+            Datamodel being processed
+
+        Returns
+        -------
+        None
+
+        """
+        is_NIR = isinstance(self, NIRDataset)
+        if is_NIR:
+            if not self.is_subarray:
+                log.info('NIR full frame data')
+                log.info('The following parameters are valid for this mode:')
+                log.info(f'use_side_ref_pixels = {self.use_side_ref_pixels}')
+                log.info(f'odd_even_columns = {self.odd_even_columns}')
+                log.info(f'side_smoothing_length = {self.side_smoothing_length}')
+                log.info(f'side_gain = {self.side_gain}')
+                log.info('The following parameter is not applicable and is ignored:')
+                log.info(f'odd_even_rows = {self.odd_even_rows}')
+            else:
+                log.info('NIR subarray data')
+                # Transform the pixeldq array from DMS to detector coords
+                self.DMS_to_detector_dq()
+                ngoodside = self.count_good_side_refpixels()
+                ngoodtopbottom = self.count_good_top_bottom_refpixels()
+                # Re-assign the pixeldq array since we transformed it to detector space
+                # and we don't want to do it again
+                self.pixeldq = self.get_pixeldq()
+                is_4amp = False
+                if self.noutputs == 4:
+                    is_4amp = True
+                if is_4amp:
+                    log.info('4 readout amplifiers used')
+                    if (ngoodside + ngoodtopbottom) == 0:
+                        log.info('No valid reference pixels.  This step will have no effect')
+                    else:
+                        log.info('The following parameters are valid for this mode:')
+                        if ngoodtopbottom > 0:
+                            log.info(f'odd_even_columns = {self.odd_even_columns}')
+                        if ngoodside > 0:
+                            log.info(f'use_side_ref_pixels = {self.use_side_ref_pixels}')
+                            log.info(f'side_smoothing_length = {self.side_smoothing_length}')
+                            log.info(f'side_gain = {self.side_gain}')
+                        log.info('The following parameters are not applicable and are ignored')
+                        if ngoodtopbottom == 0:
+                            log.info(f'odd_even_columns = {self.odd_even_columns}')
+                        if ngoodside == 0:
+                            log.info(f'use_side_ref_pixels = {self.use_side_ref_pixels}')
+                            log.info(f'side_smoothing_length = {self.side_smoothing_length}')
+                            log.info(f'side_gain = {self.side_gain}')
+                        log.info(f'odd_even_rows = {self.odd_even_rows}')
+                else:
+                    log.info('Single readout amplifier used')
+                    if ngoodtopbottom == 0:
+                        log.info('No valid reference pixels.  This step wil have no effect')
+                    else:
+                        log.info('The following parameter is valid for this mode:')
+                        log.info(f'odd_even_columns = {self.odd_even_columns}')
+                        log.info('The following parameters are not applicable and are ignored:')
+                        log.info(f'use_side_ref_pixels = {self.use_side_ref_pixels}')
+                        log.info(f'side_smoothing_length = {self.side_smoothing_length}')
+                        log.info(f'side_gain = {self.side_gain}')
+                        log.info(f'odd_even_rows = {self.odd_even_rows}')
+        else:
+            if not self.is_subarray:
+                log.info('MIRI full frame data')
+                log.info('The following parameter is valid for this mode:')
+                log.info(f'odd_even_rows = {self.odd_even_rows}')
+                log.info('The following parameters are not applicable and are ignored:')
+                log.info(f'use_side_ref_pixels = {self.use_side_ref_pixels}')
+                log.info(f'odd_even_columns = {self.odd_even_columns}')
+                log.info(f'side_smoothing_length = {self.side_smoothing_length}')
+                log.info(f'side_gain = {self.side_gain}')
+            else:
+                log.info('MIRI subarray data')
+                log.info('refpix processing skipped for this mode')
+
+    def count_good_side_refpixels(self):
+        donotuse = dqflags.pixel['DO_NOT_USE']
+        ngood = 0
+        for amplifier in 'AD':
+            rowstart, rowstop, colstart, colstop = NIR_reference_sections[amplifier]['side']
+            good = np.where(np.bitwise_and(self.pixeldq[rowstart:rowstop, colstart:colstop], donotuse) != donotuse)
+            ngood += len(good[0])
+        return ngood
+
+    def count_good_top_bottom_refpixels(self):
+        donotuse = dqflags.pixel['DO_NOT_USE']
+        ngood = 0
+        for edge in ['top', 'bottom']:
+            for amplifier in 'ABCD':
+                rowstart, rowstop, colstart, colstop = NIR_reference_sections[amplifier][edge]
+                good = np.where(np.bitwise_and(self.pixeldq[rowstart:rowstop, colstart:colstop], donotuse) != donotuse)
+                ngood += len(good[0])
+        return ngood
+
 
 class NIRDataset(Dataset):
     """Generic NIR detector Class.
@@ -1826,6 +1928,7 @@ def correct_model(input_model, odd_even_columns,
     if input_dataset is None:
         status = SUBARRAY_DOESNTFIT
         return status
+    input_dataset.log_parameters()
     reference_pixel_correction(input_dataset)
 
     return REFPIX_OK